+++ docs/CHANGELOG.md
@@ -1,8 +1,20 @@
 # Releases
 
-## v1.27.0
+## v1.28.0
 
 Released: ???
+
+Features:
+- Added option to compact sea creature catch messages [disabled by default]. It shortens double hook message and catch message that says what sea creature you caught.
+- Added option to share the location to ALL chat on Plhlegblast catch [disabled by default].
+- Added option to show caught trophy fish rarities in Odger's Trophy Fishing GUI. [disabled by default] 
+
+Bugfixes:
+- Changed Plhlegblast's rarity to legendary in the alert and Rare Catches tracker.
+
+## v1.27.0
+
+Released: 2025-02-10
 
 Features:
 - Added 2 variations of the Beach Balls and Golden Bait to the fishing profit tracker.